feat: exception delete user

+++ src/main/java/com/lsprojects/course/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.lsprojects.course.entities.User;
 import com.lsprojects.course.repositories.UserRepository;
+import com.lsprojects.course.services.exceptions.DatabaseException;
+import com.lsprojects.course.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
@@ -20,19 +23,28 @@ public class UserService {
 
 	public User findById(Long id) {
 		Optional<User> obj = userRepository.findById(id);
-		return obj.get();
+		System.out.println("ta na segunda camada de findiByid");
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return userRepository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		try {
+			if (!userRepository.existsById(id)) {
+				throw new ResourceNotFoundException(id);
+			}
+			userRepository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			System.err.println("peguei no catch error db");
+			throw new DatabaseException(e.getMessage());
+		}
 	}
-	
-	public User update (Long id, User obj) {
-		
+
+	public User update(Long id, User obj) {
+
 		User entity = userRepository.getReferenceById(id);
 		updateDate(entity, obj);
 		return userRepository.save(entity);
@@ -43,21 +55,5 @@ public class UserService {
 		entity.setName(obj.getName());
 		entity.setPhone(obj.getPhone());
 	}
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
-	
+
 }
